+++ include/PLT/TerminalPaper.h
@@ -808,7 +808,7 @@ public:
       return status;
    }
 
-   virtual ssize_t write(const void* buffer, size_t n) override
+   virtual int write(const void* buffer, size_t n) override
    {
       const uint8_t* ptr = static_cast<const uint8_t*>(buffer);
 
@@ -820,7 +820,7 @@ public:
       return n;
    }
 
-   virtual ssize_t read(void* buffer_, size_t n) override
+   virtual int read(void* buffer_, size_t n) override
    {
       uint8_t* buffer = static_cast<uint8_t*>(buffer_);
